refactor(voice-engine): extract Gemini Live event loop into native_session module
